requirements.txt
meraki>=1.34.0
pyyaml>=6.0
pandas>=1.5.0
 